Better types for errors, limits and offsets

+++ src/datastore/traits.rs
@@ -24,9 +24,9 @@ pub trait Transaction<I: Id> {
 	fn set_edge(&self, models::Edge<I>) -> Result<(), Error>;
 	fn delete_edge(&self, I, String, I) -> Result<(), Error>;
 
-	fn get_edge_count(&self, I, String) -> Result<i64, Error>;
-	fn get_edge_range(&self, I, String, i64, i32) -> Result<Vec<models::Edge<I>>, Error>;
-	fn get_edge_time_range(&self, I, String, Option<NaiveDateTime>, Option<NaiveDateTime>, i32) -> Result<Vec<models::Edge<I>>, Error>;
+	fn get_edge_count(&self, I, String) -> Result<u64, Error>;
+	fn get_edge_range(&self, I, String, u64, u16) -> Result<Vec<models::Edge<I>>, Error>;
+	fn get_edge_time_range(&self, I, String, Option<NaiveDateTime>, Option<NaiveDateTime>, u16) -> Result<Vec<models::Edge<I>>, Error>;
 
 	fn get_global_metadata(&self, String) -> Result<JsonValue, Error>;
 	fn set_global_metadata(&self, String, JsonValue) -> Result<(), Error>;
